Add generic darken on 'Create Account' hover
The 'Create Account' button slightly darkens by 8% on hover.
The way this is implemented works generically no matter the theme
color, without the need to define 3 new colors, and uses no JS.
It defines the button color in hsl() with a '--darken' variable
substracted on the lightness. The default value of darken is 0%,
and 8% on hover.