Add form validations
Every input now checks for a minimum validation requirement.
Some are more strict, like the password, which has a strength meter
and requires at least the password to be mildly strong, others
are more permisive, like the name and surname, which just require
the field not to be empty. The inputs don't show feedback the first time
the user is typing, not to be distractive. The feedback is shown on blur,
with a green border if correct, or red border and message if incorrect.
Once the user has blured the input once, then a little bit more feedback
is shown on real time to provide a bit more help if needed. The password
field also checks for matching passwords.